requirements.txt
matplotlib==3.3.2
pandas==1.1.3
streamlit==0.72.0
seaborn==0.11.0
numpy==1.19.2
Pillow==8.0.1
